Rolled back trunk to r755 to prepare for merging the 16-bit branch.
git-svn-id: svn://vcs.exim.org/pcre/code/trunk@835 2f5784b3-3f2a-0410-8824-cb99058d5e15

+++ pcre_dfa_exec.c
@@ -2781,7 +2781,7 @@ for (;;)
             {
             const uschar *p = ptr;
             const uschar *pp = local_ptr;
-            charcount = (int)(pp - p);
+            charcount = pp - p;
             while (p < pp) if ((*p++ & 0xc0) == 0x80) charcount--;
             ADD_NEW_DATA(-next_state_offset, 0, (charcount - 1));
             }
